Setup PeopleFragment to retrieve data using Volley and pass it on to the adapter
Whenever we are dealing with a dynamic list view, we need some kind of an adapter. Since we are using RecyclerView to display our list of People cards, we need to define and set a RecyclerView.Adapter for our fragment. This adapter basically determines how data is populated on to each list item and gives us more control over each item. For example, we can change color of a card based on whether it is at an odd or even postion. Try it out!